# docs/swagger.yaml
basePath: /
definitions:
  model.Article:
    properties:
      article_content:
        type: string
      article_name:
        type: string
      contentID:
        description: 关联到 content 表
        type: integer
      createTime:
        type: integer
      id:
        type: integer
      updateTime:
        type: integer
      userID:
        description: 关联到 user 表
        type: integer
    type: object
  model.CasbinModel:
    properties:
      method:
        type: string
      path:
        type: string
      ptype:
        type: string
      role_name:
        type: string
    type: object
  model.Comment:
    properties:
      articleID:
        type: integer
      comment_content:
        type: string
      createTime:
        type: integer
      id:
        type: integer
      is_thumbs_up:
        type: string
      parentID:
        type: integer
      updateTime:
        type: integer
      userID:
        type: integer
    type: object
  model.Content:
    properties:
      content_name:
        type: string
      createTime:
        type: integer
      id:
        type: integer
      introduce:
        type: string
      updateTime:
        type: integer
      userID:
        description: 关联到 user 表
        type: integer
    type: object
  model.User:
    properties:
      createTime:
        type: integer
      id:
        type: integer
      mainPageUrl:
        type: string
      nick_name:
        type: string
      password:
        type: string
      role:
        type: string
      updateTime:
        type: integer
      user_name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a blog
  title: Blog Swagger
  version: "1.0"
paths:
  /article:
    get:
      consumes:
      - application/json
      description: get all article
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: get_all_article
      tags:
      - Article
  /article/{content_name}:
    get:
      consumes:
      - application/json
      description: create article
      parameters:
      - description: 文章
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Article'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: create_article
      tags:
      - Article
  /article/{content_name}/{article_name}:
    delete:
      consumes:
      - application/json
      description: delete article
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: delete_article
      tags:
      - Article
    put:
      consumes:
      - application/json
      description: update article
      parameters:
      - description: 文章
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Article'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: update_article
      tags:
      - Article
  /article/{user_name}/{content_name}/{article_name}:
    get:
      consumes:
      - application/json
      description: get article
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: get article
      tags:
      - Article
  /comment/{article_id}:
    get:
      consumes:
      - application/json
      description: get all comment
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: get_all_comment
      tags:
      - Comment
  /comment/{comment_id}:
    delete:
      consumes:
      - application/json
      description: delete comment
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: delete comment
      tags:
      - Comment
  /comment/{user_name}/{article_id}/{parent_id}:
    post:
      consumes:
      - application/json
      description: create comment
      parameters:
      - description: 评论
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Comment'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: create_comment
      tags:
      - Comment
  /content:
    delete:
      consumes:
      - application/json
      description: delete content
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: delete content
      tags:
      - Content
    get:
      consumes:
      - application/json
      description: get all content
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: get_all_content
      tags:
      - Content
    post:
      consumes:
      - application/json
      description: create a content
      parameters:
      - description: 分类
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Content'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: create content
      tags:
      - Content
    put:
      consumes:
      - application/json
      description: update content
      parameters:
      - description: 分类
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Content'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: update content
      tags:
      - Content
  /log:
    get:
      consumes:
      - application/json
      description: get all log
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: ping log
      tags:
      - Log
  /log/{log_id}:
    delete:
      consumes:
      - application/json
      description: delete a log
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: ping log
      tags:
      - Log
  /login:
    post:
      consumes:
      - application/json
      description: login
      parameters:
      - description: 用户
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: login
      tags:
      - User
  /register:
    post:
      consumes:
      - application/json
      description: register
      parameters:
      - description: 用户
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: register
      tags:
      - User
  /role:
    delete:
      consumes:
      - application/json
      description: delete roleAuth
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: delete roleAuth
      tags:
      - Role
    get:
      consumes:
      - application/json
      description: get all roleAuth
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: get_all_roleAuth
      tags:
      - Role
    post:
      consumes:
      - application/json
      description: create roleAuth
      parameters:
      - description: 角色权限
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.CasbinModel'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: create roleAuth
      tags:
      - Role
  /role/{role_name}:
    get:
      consumes:
      - application/json
      description: get roleAuth
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: get_roleAuth
      tags:
      - Role
  /user/{user_name}:
    delete:
      consumes:
      - application/json
      description: delete user
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: delete_user
      tags:
      - User
    put:
      consumes:
      - application/json
      description: update user
      parameters:
      - description: 用户
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: update_user
      tags:
      - User
schemes:
- http
- https
swagger: "2.0"